requirements.txt
streamlit
mysql-connector-python
pandas
numpy
scikit-learn
plotly
python-dotenv
psycopg2-binary




